+++ spec/integrations/seek/back_with_reset_and_mark_again_spec.rb
@@ -21,7 +21,7 @@ class Consumer < Karafka::BaseConsumer
       next unless message.offset == 9
 
       @seeked = true
-      seek(1, reset_offset: true)
+      seek(1)
 
       return
     end